+++ src/common/debug.ml
@@ -76,9 +76,12 @@ let new_logger : char -> string -> string -> logger = fun key name desc ->
 let logger_hndl = new_logger 'h' "hndl" "command handling"
 let log_hndl = logger_hndl.logger
 
+(** To print time data. *)
+let print_time = ref false
+
 (** [time_of f x] computes [f x] and the time for computing it. *)
 let time_of : logger -> (unit -> 'b) -> 'b = fun lg f ->
-  if !log_enabled then
+  if !print_time && !log_enabled then
       let t0 = Sys.time() in
       try let y = f () in lg.logger "%f" (Sys.time() -. t0); y
       with e -> lg.logger "%f" (Sys.time() -. t0); raise e